[delete] tryGame 메소드 삭제
코드의 가독성을 해치고 불필요한 메소드이기 때문에 삭제하고 기존 코드 수정
