+++ BaseballGame_weak2/BaseballGame_weak2/Lv5/PlayData.swift
@@ -1,0 +1,15 @@
+//
+//  PlayData.swift
+//  BaseballGame_weak2
+//
+//  Created by 장상경 on 11/6/24.
+//
+
+import Foundation
+
+struct PlayData {
+    var playCount: Int = 0
+    var tryCount: Int = 0
+    
+    
+}
